refactor: 커넥션 만드는 객체 분리
